Extracted migrations and seed to Data project
