+++ TodoList.Core/Interfaces/ITodoItemService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using TodoList.Core.Models;
 
@@ -6,17 +7,17 @@ namespace TodoList.Core.Interfaces
 {
     public interface ITodoItemService
     {
-        Task<TodoItem[]> GetIncompleteItemsAsync(ApplicationUser currentUser);
-        Task<TodoItem[]> GetCompleteItemsAsync(ApplicationUser currentUser);
-        Task<TodoItem[]> GetItemsByTagAsync(ApplicationUser currentUser, string tag);
+        Task<IEnumerable<TodoItem>> GetIncompleteItemsAsync(ApplicationUser currentUser);
+        Task<IEnumerable<TodoItem>> GetCompleteItemsAsync(ApplicationUser currentUser);
+        Task<IEnumerable<TodoItem>> GetItemsByTagAsync(ApplicationUser currentUser, string tag);
         Task<bool> AddItemAsync(TodoItem todo, ApplicationUser currentUser);
         Task<bool> UpdateDoneAsync(Guid id, ApplicationUser currentUser);
         bool Exists(Guid id);
         Task<bool> UpdateTodoAsync(TodoItem todo, ApplicationUser currentUser);
         Task<TodoItem> GetItemAsync(Guid id);
         Task<bool> DeleteTodoAsync(Guid id, ApplicationUser currentUser);
-        Task<TodoItem[]> GetRecentlyAddedItemsAsync(ApplicationUser currentUser);
-        Task<TodoItem[]> GetDueTo2DaysItems(ApplicationUser user);
+        Task<IEnumerable<TodoItem>> GetRecentlyAddedItemsAsync(ApplicationUser currentUser);
+        Task<IEnumerable<TodoItem>> GetDueTo2DaysItems(ApplicationUser user);
         Task<bool> SaveFileAsync(Guid todoId, ApplicationUser currentUser, string path, long size);
     }
 }
