+++ TodoList.UnitTests/Controllers/ManageUsersControllerTest.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Moq;
 using System.Threading.Tasks;
+using TodoList.Core;
 using TodoList.Core.Models;
 using TodoList.UnitTests.Resources;
 using TodoList.Web;
