+++ TodoList.UnitTests/Models/TodoItemTest.cs
@@ -22,8 +22,12 @@ namespace TodoList.UnitTests.Models
                 Content = "Clean the kitchen.",
                 Added = new Instant(),
                 DueTo = new Instant(),
+#pragma warning disable CS0618 // 'TodoItem.AddedDateTime' is obsolete: 'Property only used for EF-serialization purposes'
                 AddedDateTime = DateTime.Now,
+#pragma warning restore CS0618 // 'TodoItem.AddedDateTime' is obsolete: 'Property only used for EF-serialization purposes'
+#pragma warning disable CS0618 // 'TodoItem.DuetoDateTime' is obsolete: 'Property only used for EF-serialization purposes'
                 DuetoDateTime = DateTime.Now,
+#pragma warning restore CS0618 // 'TodoItem.DuetoDateTime' is obsolete: 'Property only used for EF-serialization purposes'
                 Done = false,
                 Tags = new[] { "test" },
                 UserId = Guid.NewGuid().ToString(),
@@ -53,8 +57,12 @@ namespace TodoList.UnitTests.Models
                 Content = "Clean the kitchen.",
                 Added = new Instant(),
                 DueTo = new Instant(),
+#pragma warning disable CS0618 // 'TodoItem.AddedDateTime' is obsolete: 'Property only used for EF-serialization purposes'
                 AddedDateTime = DateTime.Now,
+#pragma warning restore CS0618 // 'TodoItem.AddedDateTime' is obsolete: 'Property only used for EF-serialization purposes'
+#pragma warning disable CS0618 // 'TodoItem.DuetoDateTime' is obsolete: 'Property only used for EF-serialization purposes'
                 DuetoDateTime = DateTime.Now,
+#pragma warning restore CS0618 // 'TodoItem.DuetoDateTime' is obsolete: 'Property only used for EF-serialization purposes'
                 Done = false,
                 Tags = new[] { "test" },
                 UserId = null, // Required
