--- TodoList.Web/Constants.cs
@@ -1,8 +1,0 @@
-﻿namespace TodoList.Web
-{
-    public static class Constants
-    {
-        public const string AdministratorRole = "Administrator";
-        public const string UserRole = "User";
-    }
-}
