+++ TodoList.Web/Controllers/AccountController.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Logging;
 using System;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using TodoList.Core;
 using TodoList.Core.Extensions;
 using TodoList.Core.Models;
 using TodoList.Web.Extensions;
