+++ TodoList.Web/Controllers/ManageUsersController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using System.Linq;
 using System.Threading.Tasks;
+using TodoList.Core;
 using TodoList.Core.Models;
 using TodoList.Web.Models;
