+++ TodoList.Web/Views/Shared/_AdminActionsPartial.cshtml
@@ -1,4 +1,5 @@
-﻿@using TodoList.Core.Models
+﻿@using TodoList.Core
+@using TodoList.Core.Models
 @using TodoList.Web
 
 @inject SignInManager<ApplicationUser> signInManager
