+++ TodoList.sln
@@ -10,7 +10,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TodoList.UnitTests", "TodoL
 		{FF767E44-D2F9-4B91-8920-8484CE1D980C} = {FF767E44-D2F9-4B91-8920-8484CE1D980C}
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TodoList.Core", "TodoList.Core\TodoList.Core.csproj", "{462D8089-9FA9-424D-96A8-FAB5CCBC955D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TodoList.Core", "TodoList.Core\TodoList.Core.csproj", "{462D8089-9FA9-424D-96A8-FAB5CCBC955D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TodoList.Data", "TodoList.Data\TodoList.Data.csproj", "{BB3D267C-33F6-4E4F-BC26-22B79563F5C0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -30,6 +32,10 @@ Global
 		{462D8089-9FA9-424D-96A8-FAB5CCBC955D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{462D8089-9FA9-424D-96A8-FAB5CCBC955D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{462D8089-9FA9-424D-96A8-FAB5CCBC955D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BB3D267C-33F6-4E4F-BC26-22B79563F5C0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BB3D267C-33F6-4E4F-BC26-22B79563F5C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BB3D267C-33F6-4E4F-BC26-22B79563F5C0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BB3D267C-33F6-4E4F-BC26-22B79563F5C0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
